services/radiology/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
python-multipart>=0.0.9
slowapi>=0.1.8
torch>=2.6.0
torchvision>=0.21.0
torchxrayvision==1.1.1
numpy>=1.24.0
Pillow>=10.0.0
pydicom==2.4.4

